re-organize command registry to be more centralized
