+++ pipeline/pipe/m/assetfile/assetfile_manager.py
@@ -11,7 +11,7 @@ import maya.cmds as mc
 import maya.mel as mel
 from env_sg import DB_Config
 from Qt import QtCore, QtWidgets
-from pipe.m.command import register_maya_command
+from pipe.m.command import maya_command
 from shared.util import get_production_path
 
 from pipe.asset.paths import BACKUP_DIRNAME, paths_for_asset
@@ -713,7 +713,7 @@ __all__ = [
 ]
 
 
-@register_maya_command(
+@maya_command(
     name="open_asset", label="Open Asset", icon="fileOpen.png", hotkey="ctrl+alt+a"
 )
 def open_asset():
