+++ pipeline/pipe/m/space_switch.py
@@ -2,11 +2,11 @@ from __future__ import annotations
 
 import maya.cmds as mc
 
-from pipe.m.command import register_maya_command
+from pipe.m.command import maya_command
 
 
 def createSpaceSwitch():
-    sel = mc.ls(sl=True)
+    sel = mc.ls(selection=True)
     sources = sel
 
     target = sel[-1]
@@ -23,13 +23,13 @@ def createSpaceSwitch():
         sourceNames.append(name)
 
     if mc.attributeQuery("spaceSwitch", node=target, exists=True):
-        mc.deleteAttr(target, at="spaceSwitch")
+        mc.deleteAttr(target, attribute="spaceSwitch")
 
     mc.addAttr(
         target,
-        ln="spaceSwitch",
-        at="enum",
-        en="default:" + colonSourceStr,
+        longName="spaceSwitch",
+        attributeType="enum",
+        enumName="default:" + colonSourceStr,
         keyable=True,
     )
 
@@ -41,40 +41,49 @@ def createSpaceSwitch():
     if not mc.objExists(grp):
         grp = mc.group(
             name=target + "_space_switch_GRP",
-            em=True,
+            empty=True,
         )
-        fix = mc.group(em=True, name=target + "_space_switch_TARG")
+        fix = mc.group(empty=True, name=target + "_space_switch_TARG")
         mc.matchTransform(fix, target)
         mc.parent(fix, grp)
-        pc = mc.parentConstraint(fix, parent, mo=True)
+        pc = mc.parentConstraint(fix, parent, maintainOffset=True)  # type: ignore
 
     if mc.listRelatives(grp, type="constraint") is not None:
         constraint = mc.listRelatives(grp, type="constraint")[0]
         mc.delete(constraint)
-    pc = mc.parentConstraint(sources, grp, mo=True)[0]
+    pc = mc.parentConstraint(sources, grp, maintainOffset=True)[0]  # type: ignore
 
-    pcTrgs = mc.parentConstraint(pc, wal=True, q=True, mo=True)
+    pcTrgs = mc.parentConstraint(
+        pc, weightAliasList=True, query=True, maintainOffset=True
+    )
 
-    defaultCond = mc.createNode("condition", n="default_COND")
+    defaultCond = mc.createNode("condition", name="default_COND")
     mc.connectAttr(target + ".spaceSwitch", defaultCond + ".firstTerm")
-    mc.setAttr(defaultCond + ".colorIfTrueR", 0)
-    mc.setAttr(defaultCond + ".colorIfFalseR", 1)
-    mc.setAttr(defaultCond + ".operation", 0)
+    mc.setAttr(defaultCond + ".colorIfTrueR", 0)  # type: ignore
+    mc.setAttr(defaultCond + ".colorIfFalseR", 1)  # type: ignore
+    mc.setAttr(defaultCond + ".operation", 0)  # type: ignore
 
     for count, source in enumerate(sources):
         print(source)
-        print(pcTrgs[count])
-        cond = mc.createNode("condition", n=pcTrgs[count] + "_COND")
+        print(pcTrgs[count])  # type: ignore
+        cond = mc.createNode("condition", name=pcTrgs[count] + "_COND")  # type: ignore
         mc.connectAttr(target + ".spaceSwitch", cond + ".firstTerm")
-        mc.setAttr(cond + ".secondTerm", count + 1)
-        mc.setAttr(cond + ".colorIfTrueR", 1)
-        mc.setAttr(cond + ".colorIfFalseR", 0)
+        mc.setAttr(cond + ".secondTerm", count + 1)  # type: ignore
+        mc.setAttr(cond + ".colorIfTrueR", 1)  # type: ignore
+        mc.setAttr(cond + ".colorIfFalseR", 0)  # type: ignore
         mc.connectAttr(defaultCond + ".outColorR", cond + ".colorIfTrueR")
-        mc.connectAttr(cond + ".outColorR", pc + "." + pcTrgs[count])
+        mc.connectAttr(cond + ".outColorR", pc + "." + pcTrgs[count])  # type: ignore
 
     mc.select(target)
 
 
-@register_maya_command(name="space_switch", label="Space Switch", category="animation")
+@maya_command(name="space_switch", label="Space Switch", category="animation")
 def run():
+    """
+    Creates a animation space switch setup. All the selected objects become switchable spaces for the final object in the selection.
+
+    For example:
+        Select the COG, chest, and hip controls, then finally the IK hand control, then run this command.
+        The IK hand will now have an attribute that lets you switch between COG, chest, and hip space.
+    """
     createSpaceSwitch()
